改善 Gradle 模块，调整部分配置
