+++ gradle/libs.versions.toml
@@ -132,7 +132,7 @@ spring = { id = "org.springframework.boot", version.ref = "spring-boot-v3" }
 spring-dependencyManagement = { id = "io.spring.dependency-management", version = "1.1.7" }
 kotlin-plugin-spring = { id = "org.jetbrains.kotlin.plugin.spring", version.ref = "kotlin" }
 
-suspendTransform = { id = "love.forte.plugin.suspend-transform", version.ref = "suspendTransform" }
+suspendTransform = { id = "love.forte.plugin.suspend-transform" } # , version.ref = "suspendTransform" } # version built in buildSrc
 
 dokka = { id = "org.jetbrains.dokka", version.ref = "dokka" }
